+++ src/landlab/components/flexure/funcs.py
@@ -107,7 +107,7 @@ def subside_point_load(load, loc, coords, params=None, out=None):
     >>> params = dict(eet=65000.0, youngs=7e10)
     >>> load = 1e9
 
-    Define a unifrom rectilinear grid.
+    Define a uniform rectilinear grid.
 
     >>> x = np.arange(0, 10000, 100.0)
     >>> y = np.arange(0, 5000, 100.0)