+++ src/landlab/components/lake_fill/lake_fill_barnes.py
@@ -129,7 +129,7 @@ class LakeMapperBarnes(Component):
     pit node IDs. Note also this component does not offer the `lake_codes` or
     `display_depression_map` options, for essentially this reason. Use
     `lake_map` instead for both. It also uses a much more Landlabbian
-    `run_one_step()` method as its driver, superceding
+    `run_one_step()` method as its driver, superseding
     DepressionFinderAndRouter's `map_depressions()`.
 
     A variety of options is provided. Flow routing is route-to-one in this
@@ -298,7 +298,7 @@ class LakeMapperBarnes(Component):
             calculated by a FlowDirector! This also means you need to instantiate
             your FlowDirector **before** you instantiate the LakeMapperBarnes.
             Note that the new topographic__steepest_slope will always be set to
-            zero, even if fill_flat=False (i.e., there is actually a miniscule
+            zero, even if fill_flat=False (i.e., there is actually a minuscule
             gradient on the new topography, which gets ignored).
         reaccumulate_flow : bool
             If True, and redirect_flow_steepest_descent is True, the run method
@@ -1796,7 +1796,7 @@ class LakeMapperBarnes(Component):
         self._runcount = next(self._runcounter)
         # First get _fill_surface in order.
         self._fill_surface[:] = self._surface  # surfaces begin identical
-        # note this is nice & efficent if _fill_surface is _surface
+        # note this is nice & efficient if _fill_surface is _surface
         # if we're doing a redirect, we're going to need to preserve this
         # initial topo, so let's do that:
         if not self._dontredirect:
@@ -2324,7 +2324,7 @@ class LakeMapperBarnes(Component):
         """A nlakes-long array of the volume of each lake. The order is the
         same as that of the keys in lake_dict, and of lake_outlets. Note that
         this calculation is performed relative to the initial surface, so is
-        only a true lake volume if the initial surface was the rock suface (not
+        only a true lake volume if the initial surface was the rock surface (not
         an earlier water level).
 
         Requires that surface and fill_surface were not the same array at